move logic into service
